requirements.txt
backcall==0.2.0
blis==0.7.3
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cymem==2.0.4
decorator==4.4.2
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
idna==2.10
importlib-metadata==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.7.0
lxml==4.6.2
murmurhash==1.0.4
nltk==3.5
numpy==1.19.4
pandas==1.1.4
parso==0.7.1
pdfminer==20191125
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.4
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycryptodome==3.9.9
Pygments==2.7.2
PyMuPDF==1.18.4
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.4
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
spacy==2.3.4
srsly==1.0.4
thinc==7.4.3
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.53.0
traitlets==5.0.5
urllib3==1.26.2
wasabi==0.8.0
wcwidth==0.2.5
zipp==3.4.0
